Make GoPackageGraph the interface between tier2 and the package graph
